requirements.txt
onnxruntime
streamlit==1.25.0
rembg==2.0.16
Pillow==9.5.0
streamlit-drawable-canvas==0.11.0